Backport dtb inclusion logic from Q
In Q boot image with header version 2 supports inclusion of
DTB image within the boot image.
QCDT legacy devices work the same way except for the dtb
size in bytes that is specified earlier in the final image.

I backported Google's logic to P to get rid of qcom stuff in build/make.
This is unneccessary on other roms at the moment but will be required in Q.